requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiosqlite==0.20.0

# Templates
jinja2==3.1.4

# HTTP client for scraping
httpx==0.28.1
beautifulsoup4==4.12.3
lxml==5.3.0

# CLI
click==8.1.7

# Environment
python-dotenv==1.0.1

# Date handling
python-dateutil==2.9.0

# Image handling
pillow==11.0.0
